Internals: Speed up V3Broken
This patch makes OpenTitan verilation with --debug-check 40% faster, and
the same with --debug --no-dump-tree 104% faster. Functionality is the
same (including when VL_LEAK_CHECKS is defined). The main changes are
the following:

- Store the V3Broken traversal state in the AstNode by stealing a byte
from what used to be unused flags. Note that the leak checks still use
an unordered_set, but that is never actually touched unless
VL_LEAK_CHECKS is defines, which we normally do not. This eliminates a
huge number of what used to be unordered_map lookups. This should all be
safe because we fail fast if encountering an unallocated node with
VL_LEAK_CHECKS, so we should not need to dereference bad pointers. This
accounts for 3/4 of the improvement.

- As both of the visitors involved in the brokenness check are const
visitors (do not modify the tree at all), and both traverse the tree in
the same order (pre-order), and as we already need to save the
enumerated node pointers in a vector, we can re-used this in the second
traversal to perform precise pre-fetching. This accounts for about 1/4
of the improvements.

As a bonus point, there should be a lot less code coverage holes now in
V3Broken, and the amount of noodle soup code has been reduced.

I have checked Verilator build with VL_LEAK_CHECKS (but there are some
pre existing leaks), and the brokenness check also seems to work base
don some manual tests.