Major refactoring of ordering for improved scheduler
There are quite a couple of changes going on in this patch. The goal is
to rework the interfaces of the passes between V3Order and V3Changed in
a way that allows us to invoke V3Order on arbitrary subsets of logic,
and create separate evaluation functions for these. This is in
preparation for the improved scheduling discussed in #3278.

Notable conceptual changes:

1. Pulled the propagation of the clocker attribute into a separate pass
(from V3Order), this will become redundant with #3278, but is still
needed now.

2. Introduced a separate event domain for static variable initializers,
which run before initial blocks, with the corresponding
`AstInitialStatic` AST node. (This has the benefit of yielding an error
when assigning to `const` variables in an initial block, which
previously used to be accepted.)

3. Added a separate `V3Sched::schedule` pass, which replaces V3Order.
V3Sched partitions the logic in the design into separate classes, and
then invokes V3Order to order the appropriate subsets to create the
evaluation functions. This both reduces working set sizes and simplifies
some internals. Specifically V3Sched creates these entities:

3a. Static initializer, initial and final blocks are simply concatenated
(i.e.: are in file order, but otherwise arbitrary) to create the
`_eval_static`, `_eval_initial`, and `_final` functions (these are
AstCFunc after `V3Sched::schedule`).

3b. The `_eval_settle` is created by ordering the all of the
combinational logic.

3c. `_eval` is created from both the combinational and clocked logic, but
not from static initializers.

4. The change detect loops are now inside `_eval_settle` and `_eval` and
are created explicitly in V3Changed rather than in V3EmitCModel. The new
`AstEval` node is used to carry the body of the evaluation loops from
V3Sched through to V3Changed, where the optional change detect loop is
added. Each `AstEval` gets its own change detect functions (they ware
shared previously), as they might have different subsets of variables
they care about.

5. The change detect functions are split in two: `_change_snapshot`,
which used to be in the `finalp()` section of the old `_change_request`
function and copies the current state of circular variables into the
"previous value" buffers. `_change_check`, which computes if we need to
loop again. The eval loop body is then roughly:

```
bool changeReq = true;
while (changeReq) (
    _change_snapshot(); // Save current state of circular variables
    _eval_body();
    changeReq = _change_check();
}
```

As mentioned, the change detect loop is now inside `_eval`. If no change
detect loop is needed, then `_eval` simply contains the ordered code. If
a change detect function is needed, an intermediate `_eval_body` function
is created to hold the body, as shown above, and `_eval` holds the loop
above.

This patch is mostly refactoring. The known behaviour differences are:
- As mentioned in 2. above, assignments to `const` variables in initial
blocks should now error.
- When the iteration limit in the settle loop is reached, debug prints
are not turned on automatically for the last change detect invocation.
This is mostly due to laziness at this point as it requires some text
bashing required.

Apart from the above, ordering of static initializers, initial and final
blocks might differ from previous versions, but is compatible with the
standard.

The patch should be performance neutral.