Add representation of complex always blocks to DFG
This patch adds DfgAlways, which is a vertex that represents a whole,
arbitrarily complex combinational always block in the DfgGraph.

Implementing this requires computing the variables live at block entry
(variables read by the block), so there is a new ControlFlowGraph
data structure and a classical data-flow analysis based live variable
analysis to do that at the variable level (as opposed to bit/element
level). The CFG ill be used further in later patches as well.

The actual CFG construction and live variable analysis is best effort,
and might fail for currently unhandled constructs or data types. This
can be extended later.

Once the always blocks have been converted into DfgAlway, the only thing
we can actually do with them at this point is delete them from the
graph. There will be more transformations later.